Refactored:
- Moved some codes to reduce the ammound of folders.
   - /error => /
   - /net/connection => /net
- Refactored some documentation comments to be more correct.
- Updated README.
- Removed some public types which should supose to be public.
