+++ README.md
@@ -15,10 +15,26 @@
 [s6]: https://tokei.rs/b1/github/amethyst/laminar?category=code
 [s7]: https://codecov.io/gh/amethyst/laminar/branch/master/graphs/badge.svg
 
-This library implements some TCP-like features on top of an UDP-socket. 
-It will provide a lightweight message-based interface with certain guarantees like reliability, fragmentation congestion monitoring.
+Laminar is a semi-reliable UDP-based protocol for multiplayer games. This library implements wrappers around the UDP-protocol,
+and provides a lightweight, message-based interface which provides certain guarantees like reliability and ordering.
 
-Laminar was designed to be used in the [Amethyst][amethyst] game engine and is loosely based on articles from [Gaffer on Games](https://gafferongames.com/).
+Laminar was designed to be used within the [Amethyst][amethyst] game engine but is usable without it.
+
+[amethyst]: https://github.com/amethyst/amethyst
+
+# Concepts
+
+This library is loosely based off of [Gaffer on Games][gog] and shares features similar as RakNet, Steam Socket, netcode.io.
+The idea is to provide an in rust written, low-level UDP-protocol which supports the use of cases of video games that require multilayer features.
+The library itself provides a few low-level types of packets that provide different types of guarantees. The most
+basic are unreliable and reliable packets. Also ordering, sequencing can be done on multiple streams.
+For more information, read the projects [README.md][readme], [book][book], [docs][docs] or [examples][examples].
+
+[gog]: https://gafferongames.com/
+[readme]: https://github.com/amethyst/laminar/blob/master/README.md
+[book]: https://github.com/amethyst/laminar/tree/master/docs/md_book
+[docs]: https://docs.rs/laminar/
+[examples]: https://github.com/amethyst/laminar/tree/master/examples
 
 [amethyst]: https://github.com/amethyst/amethyst
 
@@ -35,16 +51,15 @@ Laminar was designed to be used in the [Amethyst][amethyst] game engine and is l
 ## Features
 These are the features this crate provides:
 
-- UDP-based protocol
-- Connection tracking
+- UDP-based Protocol
+- Connection Tracking
 - Automatic Fragmentation
-- Unreliable and Reliable packets
-- Protocol versioning
-- RTT estimation
+- Packets Types: Unreliable and Sequenced, Reliable and Unordered, Sequenced and Ordered.
+- Arranging Streams
+- Protocol Versioning
+- RTT Estimation
 - Link conditioner to simulate packet loss and latency
-- Well tested by integration tests and unit tests
-- Good error handling
-- Benchmarks
+- Well-tested by integration and unit tests
 
 ## Getting Stated
 Add the laminar package to your `Cargo.toml` file.
@@ -62,8 +77,7 @@ laminar = "0.1"
 - [Contributing](https://github.com/amethyst/laminar/blob/master/docs/CONTRIBUTING)
 
 ## Examples
-These are some basic examples demonstrating how to use this crate. 
-Please checkout our [examples](https://github.com/amethyst/laminar/tree/master/examples) for more.
+Please check out our [examples](https://github.com/amethyst/laminar/tree/master/examples) for more information.
 
 ### UDP API | [see more](https://github.com/amethyst/laminar/blob/master/examples/udp.rs)
 This is an example of how to use the UDP API.
@@ -74,7 +88,7 @@ _Send packets_
 use laminar::{DeliveryMethod, Packet};
 use laminar::net::{UdpSocket, NetworkConfig};
 
-// Create the necessarily config, you can edit it or just use the default.
+// Create the necessary config, you can edit it or just use the default.
 let config = NetworkConfig::default();
 
 // Setup an udp socket and bind it to the client address.
@@ -133,15 +147,12 @@ match result {
 - [Lucio Franco](https://github.com/LucioFranco)
 - [Fletcher Haynes](https://github.com/fhaynes)
 - [Timon Post](https://github.com/TimonPost)
-- [Justin LeFebvre](https://github.com/jstnlef)
-
-We want to give credit to [gaffer on games](https://gafferongames.com/) as we have used his guide to building a game networking protocol to build this library. 
+- [Justin LeFebvre](https://github.com/jstnlef) 
 
 ## Note
 
-This library is not fully stable yet. 
-Although version 0.1.0 is released we might have to change some of the existing API.  
-Laminar is used in [Amethyst-Network](https://github.com/amethyst/amethyst/tree/master/amethyst_network), you could give that a look if you want to see some more advanced use-cases.
+This library is not fully stable yet, and there may be breaking changes to the API.
+For more advanced examples of using laminar, you can check out the [Amethyst-Network](https://github.com/amethyst/amethyst/tree/master/amethyst_network) crate.
 
 ## Contribution
