+++ src/infrastructure.rs
@@ -1,9 +1,16 @@
+//! This module provides the logic around the processing of the packet.
+//! Like ordering, sequencing, controlling congestion, fragmentation, and packet acknowledgment.
+
 mod acknowlegement;
 mod congestion;
+mod external_ack;
 mod fragmenter;
+mod local_ack;
 
 pub mod arranging;
 
 pub use self::acknowlegement::AcknowledgementHandler;
 pub use self::congestion::CongestionHandler;
+pub use self::external_ack::ExternalAcks;
 pub use self::fragmenter::Fragmentation;
+pub use self::local_ack::LocalAckRecord;
