+++ src/net/constants.rs
@@ -1,18 +1,19 @@
-/// Fragment header size
+/// The size of the fragment header.
 pub const FRAGMENT_HEADER_SIZE: u8 = 4;
-/// Acked packet header size
+/// The size of the acknowledgement header.
 pub const ACKED_PACKET_HEADER: u8 = 8;
-/// Arranging packet header size
+/// The size of the arranging header.
 pub const ARRANGING_PACKET_HEADER: u8 = 3;
-/// Standard header size
+/// The size of the standard header.
 pub const STANDARD_HEADER_SIZE: u8 = 5;
+/// The ordering stream that will be used to order on if there is not ordering stream specified.
 pub const DEFAULT_ORDERING_STREAM: u8 = 255;
+/// The sequencing stream that will be used to sequence packets on if there is not sequencing stream specified.
 pub const DEFAULT_SEQUENCING_STREAM: u8 = 255;
-/// Default max number of fragments to size
+/// Default maximal number of fragments to size.
 pub const MAX_FRAGMENTS_DEFAULT: u16 = 16;
-/// Default max size of each fragment
+/// Default maximal size of each fragment.
 pub const FRAGMENT_SIZE_DEFAULT: u16 = 1024;
-
 /// Maximum transmission unit of the payload.
 ///
 /// Derived from ethernet_mtu - ipv6_header_size - udp_header_size - packet header size
