+++ src/net/events.rs
@@ -1,16 +1,15 @@
 use crate::packet::Packet;
 use std::net::SocketAddr;
 
-/// Events which will be pushed through the event_receiver returned by `Socket::bind`.
+/// Events that can occur in `laminar` and that will be pushed through the `event_receiver` returned by `Socket::bind`.
 #[derive(Debug, PartialEq)]
 pub enum SocketEvent {
-    /// A packet has been received from a client.
+    /// A packet was received from a client.
     Packet(Packet),
-    /// A new client connects. Clients are uniquely identified by the ip:port combination at this layer.
+    /// A new client connected.
+    /// Clients are uniquely identified by the ip:port combination at this layer.
     Connect(SocketAddr),
-    /// A client disconnects. This is generated from the server-side intentionally disconnecting a client,
-    /// or it could be from the client disconnecting.
-    Disconnect(SocketAddr),
-    /// This is generated if the server has not seen traffic from a client after a configurable amount of time.
+    /// The client has been idling for a configurable amount of time.
+    /// You can control the timeout in the config.
     Timeout(SocketAddr),
 }
